Refactor XML validation into a class
Previously, we only had a single static function that read the XML
schema on every XML validation. This change refactors the validation
into a separate class so that the schema does not have to be reloaded.

This is practical for validating a large number of XML samples.

Finally, we implement a table-driven test of the validation based on
the XML files directly extracted from the AASXs obtained from
http://admin-shell-io.com/samples/.